Add discounted price calculation to membership model
A couple of new fields, 'monthly_cost_original' and 'annual_cost_original' have been added to the 'costs' object in the 'membership' model. These fields are used to calculate and store the original price before discounts. A new condition is now checked for each membership type to calculate and set its original price if the membership type is 'country'.