Add dynamic title support and update membership model
Updated routing to include dynamic title data for each page. The title data is then observed by AppComponent and set as the current page title. In the membership model, additional getter and setter methods are added for the nanny and past_member fields. Moreover, the application-form component is revised, and able to handle incoming query parameters to establish a new membership object.